refactored the chart options #126
