+++ local-config.yml
@@ -51,6 +51,6 @@ utils:
     mean: 0
   services-state:
     ping: passive
-    flood-ping: passive
+    flood-ping: active
     jitter: passive
     req-res-delay-or-monitoring: passive
